split main into separate functions

+++ pcgpmem.c
@@ -1,13 +1,13 @@
 #include <assert.h>
 #include "pcgpmem.h"
 
-void arenaInit(Arena *a, void *mem, size_t size) {
+void arenaInit(Arena* a, void* mem, size_t size) {
 	a->start = mem;
-	a->end = a->start + size;
+	a->end = (uintptr_t)a->start + size;
 	a->ptr = a->start;
 }
 
-void *arenaAlloc(Arena *a, size_t size) {
+void* arenaAlloc(Arena* a, size_t size) {
 	uintptr_t ptr = (uintptr_t)a->ptr;
 	uintptr_t aptr = (uintptr_t)ptr;
 	uintptr_t mod = ptr & (PCGPMEM_ALIGN - 1);
@@ -16,10 +16,10 @@ void *arenaAlloc(Arena *a, size_t size) {
 	}
 	ptr = aptr + size;
 	assert(ptr <= (uintptr_t)a->end);
-	a->ptr = (void *)ptr;
-	return (void *)aptr;
+	a->ptr = (void*)ptr;
+	return (void*)aptr;
 }
 
-void arenaFree(Arena *a) {
+void arenaFree(Arena* a) {
 	a->ptr = a->start;
 }
